Update Spaceplane Plus to v1.3 for KSP 0.24.2
This needs https://github.com/kmc/kmc/pull/426 to work.

Made module manager a pre-requisite. Created a custom postrequisite for the Firespitter DLL. In the original download, an outdated Firespitter DLL was situated in a small directory structure, which Kmc could not re-create. https://github.com/kmc/kmc/pull/426 adds directory creation to Kmc, which is needed to install the Firespitter DLL as it is in the zip from SpaceplanePlus.